refactor: Changes of file structure
