+++ labirint/package.json
@@ -11,6 +11,7 @@
     "pixi.js-legacy": "^5.3.7",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
     "web-vitals": "^0.2.4"
   },
