+++ labirint/src/App.css
@@ -1,3 +1,5 @@
+@import './components/css/reset.css';
+
 .App {
   text-align: center;
 }
