+++ labirint/src/components/_Cover/cover.module.scss
@@ -1,4 +1,5 @@
 .cover {
+  display: block;
   z-index: 500;
   position: fixed;
   top: 7vh;
