+++ labirint/src/components/_Game/Game.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import styleObj from './game.module.scss'
+
+function Game() {
+  return (
+    <div className={styleObj.game}></div>   
+  )  
+}
+
+export default Game;
