+++ labirint/src/components/_Game/game.module.scss
@@ -1,0 +1,9 @@
+.game {
+  background: radial-gradient(farthest-corner at 50vw 50vh, rgba(2, 0, 36, 0) 14%, rgba(52, 40, 28, 1) 30%, rgba(8, 2, 0, 1) 100%);
+  z-index: 500;
+  position: fixed;
+  top: 7vh;
+  left: 0;
+  width: 100vw;
+  height: 86vh;
+}
