+++ labirint/src/components/_Header/header.module.scss
@@ -1,4 +1,4 @@
-@import 'shadow-hover';
+@import '../sass/shadow-hover.scss';
 
 .header {
   position: fixed;
