+++ labirint/src/components/css/reset.css
@@ -1,0 +1,49 @@
+*,
+*::before,
+*::after {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  min-height: 100vh;
+  scroll-behavior: smooth;
+  text-rendering: optimizeSpeed;
+  line-height: 1.5;
+}
+
+a {
+  color: black;
+  text-decoration: none;
+  list-style: none;
+  border: none;
+  background: none;
+}
+
+a:link, a:visited, a:focus {
+  color: black;
+  background: none;
+}
+
+a:hover {
+  background: none;
+}
+
+a:active {
+  background: none;
+  color: rgb(83, 25, 25);
+  text-decoration: underline;
+}
+
+img {
+  max-width: 100%;
+  display: block;
+}
+
+input,
+button,
+textarea,
+select {
+  font: inherit;
+}
